requirements.txt
protobuf<=3.20.1
torch
transformers
datasets
wandb
evaluate
numpy
fire
tabulate==0.9.0
pillow==10.0.1
spacy
diffusers==0.21.4
accelerate
torchinfo
lpips
webdataset
einops
omegaconf